third_party/blink/web_tests/fast/dom/MutationObserver/create-during-delivery-expected.txt

Test that MutationObservers created during delivery must wait for the next loop.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS order.length is 4
PASS order[0] is 1
PASS order[1] is 2
PASS order[2] is 1
PASS order[3] is 3
PASS successfullyParsed is true

TEST COMPLETE
